+++ GUI/GUIRadioButton.cpp
@@ -122,7 +122,7 @@ void GUIRadioButton::ChangeSkin(GUISkin *Skin)
 void GUIRadioButton::BuildBitmap(void)
 {
     string Filename;
-    Uint32 ColorIndex = 0;
+    unsigned long ColorIndex = 0;
     int Values[4];
 
     // Load the image
